requirements.txt
Django==1.6.7
MySQL-python==1.2.3
django-admin-tools==0.5.2
djangorestframework==2.4.1
feedparser==5.1.3
gunicorn
django-filter==0.9.1
South==1.0.1